Pure commit matching (no more reviewing file summaries to filter).
It should improve performance, as the review can generally only
limit the potential relevant files, and the max files retrieved
is capped by match-strength anyway.

The reduction of relevant files by 1 or 2 is outweighed by
eliminating the extra round of inference (including extra time),
and eliminating risk of dropping a relevant file in the process.